requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
selenium>=4.15.2
lxml>=4.9.3
pandas>=2.1.3
python-dotenv>=1.0.0
aiohttp>=3.9.1
fake-useragent>=1.4.0
pytest>=7.4.3
black>=23.11.0
flake8>=6.1.0
isort>=5.12.0
rich>=13.7.0
loguru>=0.7.2
tqdm>=4.66.1 